+++ imports/framework/Forms/DetailsForm/Actions/DetailsForm.Actions.js
@@ -1,7 +1,7 @@
 import { Template } from 'meteor/templating';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 
-import { getEntityTranslation } from '/imports/framework/Helpers';
+import { getEntityTranslation } from '/imports/framework/Helpers/Helpers';
 
 import './DetailsForm.Actions.jade';
