+++ imports/framework/Forms/DetailsForm/Dropdown/DetailsForm.Dropdown.js
@@ -3,7 +3,7 @@ import { TAPi18n } from 'meteor/tap:i18n';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 
 import { getValue, getKey } from '/imports/framework/Forms/DetailsForm/DetailsForm.Helpers';
-import { getEntityTranslation } from '/imports/framework/Helpers';
+import { getEntityTranslation } from '/imports/framework/Helpers/Helpers';
 
 import './DetailsForm.Dropdown.jade';
