+++ imports/framework/Forms/DetailsForm/EntityArray/DetailsForm.Array.Entity.js
@@ -4,7 +4,7 @@ import { TAPi18n } from 'meteor/tap:i18n';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 
 import { getValue, getKey, isType } from '/imports/framework/Forms/DetailsForm/DetailsForm.Helpers';
-import { getEntityTranslation } from '/imports/framework/Helpers';
+import { getEntityTranslation } from '/imports/framework/Helpers/Helpers';
 
 import './DetailsForm.Array.Entity.jade';
