+++ imports/framework/Forms/InsertForm/Textbox/InsertForm.Textbox.js
@@ -3,8 +3,8 @@ import { Template } from 'meteor/templating';
 
 import './InsertForm.Textbox.jade';
 
-import { getKey, getTitle, getEntityTranslation } from '/imports/framework/Helpers';
-import { hasError, getErrorClass, getEntityErrorTranslation } from '/imports/framework/Helpers.Error';
+import { getKey, getTitle, getEntityTranslation } from '/imports/framework/Helpers/Helpers';
+import { hasError, getErrorClass, getEntityErrorTranslation } from '/imports/framework/Helpers/Helpers/Error';
 
 Template.InsertFormTextbox.helpers({
   getKey,
