+++ imports/framework/Forms/UpdateForm/Textbox/UpdateForm.Textbox.js
@@ -1,8 +1,8 @@
 import autosize from 'autosize';
 import { Template } from 'meteor/templating';
 
-import { getKey, getTitle, getEntityTranslation } from '/imports/framework/Helpers';
-import { hasError, getErrorClass, getEntityErrorTranslation } from '/imports/framework/Helpers.Error';
+import { getKey, getTitle, getEntityTranslation } from '/imports/framework/Helpers/Helpers';
+import { hasError, getErrorClass, getEntityErrorTranslation } from '/imports/framework/Helpers/Helpers/Error';
 
 import './UpdateForm.Textbox.jade';
