+++ imports/framework/Forms/UpdateForm/UpdateForm.js
@@ -4,8 +4,8 @@ import { ReactiveVar } from 'meteor/reactive-var';
 import { TAPi18n } from 'meteor/tap:i18n';
 import { FlowRouter } from 'meteor/kadira:flow-router';
 
-import { getKey, getTitle, getEntityTranslation } from '/imports/framework/Helpers';
-import { hasError, getErrorClass, getEntityErrorTranslation } from '/imports/framework/Helpers.Error';
+import { getKey, getTitle, getEntityTranslation } from '/imports/framework/Helpers/Helpers';
+import { hasError, getErrorClass, getEntityErrorTranslation } from '/imports/framework/Helpers/Helpers/Error';
 
 import './UpdateForm.jade';
 import './UpdateForm.scss';
